Защита списка лотов от XSS
