+++ index.php
@@ -1,7 +1,6 @@
 <?php
 require_once('functions.php'); //подключаем сценарий с функцией-шаблонизатором
 require_once('data.php'); //подключаем сценарий с данными
-require_once('price_format.php'); //подключаем сценарий с функцией обрабатывающей строку с ценой
 $is_auth = rand(0, 1);
 $user_name = 'Лаура'; // укажите здесь ваше имя
