+++ templates/card.php
@@ -4,7 +4,7 @@
                     </div>
                     <div class="lot__info">
                         <span class="lot__category"><?=$value['category']; ?></span>
-<h3 class="lot__title"><a class="text-link" href="pages/lot.html"><?=$value['title']; ?></a></h3>
+<h3 class="lot__title"><a class="text-link" href="pages/lot.html"><?=htmlspecialchars($value['title']); ?></a></h3>
 <div class="lot__state">
     <div class="lot__rate">
         <span class="lot__amount">Начальная цена</span>
